Fix bug that caused immediate shutdown on boot
Had added a loop to make sure the pi_running flag in the micro was
set, but the bug would allow the loop to be quit on read error of
the flag as well.  Changed the condition to fix that and added
delay to loop.
To debug this, added syslog logging of daemon events as well.