Add register version 4 with VIN (USB voltage)
Add registers related to input (USB voltage) measurement and
related features such as new auto boot on USB connect and auto
shutdown after a programmable time when USB is disconnected.

Restructured the variable table to be more compact and easier
to read, with a list of all register addresses in the table.