Added "smart" auto boot options
In addition to automatically booting when Vbat is high enough or
when Vbat and Vin are both high enough, each of these options now
has a "smart" variant that will only auto boot when the last shutdown
was due to a low voltage condition, either low Vin or low Vbat.

This makes it so you can still manually turn the device on and off
with the touch button or programmatically.