Add I2C register read redundancy
The I2C read needs to return 3 identical readings in a row to be
accepted.  The reason to add this is that I have not succeeded in
making the I2C read 100% reliable. This is because the MSP430G I2C
peripheral is not a full hardware peripheral but requires a ton
of software intervention to make it work.  It seems sometimes the
interrupt handlers are delayed enough (likely because another
interrupt was already running) that the I2C doesn't get set up
correctly in time before the master needs the next bits.