Add delay between register write and data read
In pursuit of a bug that overwrites I2C registers such as I2C
address and LED state, the current theory is that the I2C system
on the micro is interpreting an address as data to be written,
since the overwriting values tend to be 0x86 or 0x87, or the
device's 8-bit address with either read or write LSB.  So it
seems it's not detecting either the STOP or START condition, or
somehow the host is messing this up and not sending them.

Either way, to make it more likely the STOP and START are detected
correctly and/or sent correctly, I've split the register write and
read accesses into two different ioctls with a 50us delay between
them.