Fix matching read count and add variable delay
There was a off-by-one bug in the logic for how many matching
reads are needed before the data is accepted so that was fixed.
Also added a variable delay between reads, faster at first, then
with more delay.
Added a sign_extend flag to the variable table, which we use to
flag 16-bit variables we want sign-extended.