Fix bug making auto shutdown fail sometimes
There was a loop that would keep setting PI_RUNNING until it read
as 1.  This caused a conflict if AUTO_SHDN_TIME was 0 and external
power was not present when the system was booted and PI_RUNNING was
set to 1 by the daemon.  The micro would immediately decide the
voltage was not present and set PI_RUNNING back to zero and go
to shutdown state.  But the daemon would immediately set PI_RUNNING
back to 1 due to the loop, causing the micro and daemon to be out
of sync.

Things work fine without the loop.  No need to complicate things
and cause issues, so the loop was removed.